Move test folder to the root directory of the project. Add a new test for search by multiple keywords
