--- pdf_parser/tests/test_main.py
@@ -1,24 +1,0 @@
-import unittest
-
-testmodules = [
-    'pdf_parser.tests.test_tools',
-    'pdf_parser.tests.test_pdf_objects',
-    ]
-
-suite = unittest.TestSuite()
-
-for t in testmodules:
-    try:
-        # If the module defines a suite() function, call it to get the suite.
-        mod = __import__(t, globals(), locals(), ['suite'])
-        suitefn = getattr(mod, 'suite')
-        suite.addTest(suitefn())
-    except (ImportError, AttributeError):
-        # else, just load all the test cases from the module.
-        suite.addTest(unittest.defaultTestLoader.loadTestsFromName(t))
-
-unittest.TextTestRunner().run(suite)
-
-
-if __name__ == '__main__':
-    unittest.main()
